+++ abci/code/code.go
@@ -89,5 +89,6 @@ const (
 	CannotGetMasterPublicKeyFromNodeID        uint32 = 62
 	CannotGetPublicKeyFromNodeID              uint32 = 63
 	TimeOutBlockIsMustGreaterThanZero         uint32 = 64
+	RSAKeyLengthTooShort											uint32 = 65
 	UnknownError                              uint32 = 999
 )